refactor(database): split models by domain
Split monolithic models.py (1312 lines) into 8 modules:

- base.py (7) - Base class
- trading.py (163) - 4 trade/position models
- analysis.py (159) - 3 analysis/signal models
- discovery.py (133) - 3 discovery models
- monitoring.py (233) - 5 health/metrics models
- portfolio.py (206) - 6 optimization models
- metrics.py (400) - 9 performance models
- screening.py (171) - 5 screening models

All <400 lines. 100% backward compatibility via __init__.py
re-exports. All 35 ORM models imported for Base.metadata.

Fix: 704
