+++ src/database/models/__init__.py
@@ -1,0 +1,91 @@
+"""Database ORM models."""
+
+from src.database.models.analysis import (
+    AnalysisRecordORM,
+    ExecutionGraphORM,
+    SignalOutcomeORM,
+)
+from src.database.models.base import Base
+from src.database.models.discovery import (
+    ActiveDiscoveryCandidateORM,
+    DiscoveryHistoryRecordORM,
+    DiscoverySourceMetricsORM,
+)
+from src.database.models.metrics import (
+    ExecutionMetricORM,
+    MonteCarloRecordORM,
+    PaperTradingReportORM,
+    ProfilingRecordORM,
+    RiskReportRecordORM,
+    SupervisorMetricsORM,
+    TearSheetORM,
+    TradeJournalORM,
+    WorkflowExecutionMetricsORM,
+)
+from src.database.models.monitoring import (
+    CoordinatorMetricsORM,
+    DaemonMetadataORM,
+    DegradationRecordORM,
+    HealthReportORM,
+    RiskAuditORM,
+)
+from src.database.models.portfolio import (
+    CorrelationAuditRecordORM,
+    OptimizationRecordORM,
+    PeerAnalysisRecordORM,
+    RebalancingRecordORM,
+    SectorAttributionRecordORM,
+    SectorRotationRecordORM,
+)
+from src.database.models.screening import (
+    EarningsCalendarRecordORM,
+    GamePlanRecordORM,
+    PrefetchRecordORM,
+    ScoringWeightsHistoryORM,
+    ScreeningRecordORM,
+)
+from src.database.models.trading import (
+    PortfolioSnapshotORM,
+    PositionManagementActionORM,
+    PositionRecordORM,
+    TradeORM,
+)
+
+__all__ = [
+    "ActiveDiscoveryCandidateORM",
+    "AnalysisRecordORM",
+    "Base",
+    "CoordinatorMetricsORM",
+    "CorrelationAuditRecordORM",
+    "DaemonMetadataORM",
+    "DegradationRecordORM",
+    "DiscoveryHistoryRecordORM",
+    "DiscoverySourceMetricsORM",
+    "EarningsCalendarRecordORM",
+    "ExecutionGraphORM",
+    "ExecutionMetricORM",
+    "GamePlanRecordORM",
+    "HealthReportORM",
+    "MonteCarloRecordORM",
+    "OptimizationRecordORM",
+    "PaperTradingReportORM",
+    "PeerAnalysisRecordORM",
+    "PortfolioSnapshotORM",
+    "PositionManagementActionORM",
+    "PositionRecordORM",
+    "PrefetchRecordORM",
+    "ProfilingRecordORM",
+    "RebalancingRecordORM",
+    "RiskAuditORM",
+    "RiskReportRecordORM",
+    "ScoringWeightsHistoryORM",
+    "ScreeningRecordORM",
+    "SectorAttributionRecordORM",
+    "SectorRotationRecordORM",
+    "SignalOutcomeORM",
+    "SupervisorMetricsORM",
+    "TearSheetORM",
+    "TradeJournalORM",
+    "TradeORM",
+    "WorkflowExecutionMetricsORM",
+]
