+++ src/database/models/base.py
@@ -1,0 +1,7 @@
+"""Base class for ORM models."""
+
+from sqlalchemy.orm import DeclarativeBase
+
+
+class Base(DeclarativeBase):
+    """Base class for ORM models."""
